refactor: add common commands package and move sozluk command functions to it
