+++ packages/common/helpers/helpers.go
@@ -1,0 +1,43 @@
+package helpers
+
+import (
+	"fmt"
+	"io"
+	"net/http"
+	"net/url"
+)
+
+func IsCommandParamsLengthEqualToOne(params []string) bool {
+	return len(params) == 1
+}
+
+type SozlukGraphQLResponse struct {
+	Data struct {
+		Sozluk struct {
+			Term struct {
+				Title string `json:"title"`
+				Body  struct {
+					Raw string `json:"raw"`
+				} `json:"body"`
+			} `json:"term"`
+		} `json:"sozluk"`
+	} `json:"data"`
+}
+
+func FetchGraphQL(apiUrl string, query string) ([]byte, error) {
+	queryParams := url.QueryEscape(query)
+	fullURL := fmt.Sprintf("%s?query=%s", apiUrl, queryParams)
+
+	resp, err := http.Get(fullURL)
+	if err != nil {
+		return nil, err
+	}
+	defer resp.Body.Close()
+
+	body, err := io.ReadAll(resp.Body)
+	if err != nil {
+		return nil, err
+	}
+
+	return body, nil
+}
